Extract recognizer display into own class
